+++ llm_mcp_server/wizards/__init__.py
@@ -1,1 +1,2 @@
 from . import mcp_key_wizard
+from . import mcp_oauth_client_secret_wizard